+++ storage/actions/list.js
@@ -9,7 +9,7 @@ const { logger, output } = require('../utils');
 /**
  *
  */
-module.exports = async (fiber) => {
+module.exports = exports = async (fiber) => {
   logger.verbose("list ...");
   let retCode = 0;