+++ storage/actions/transfer.js
@@ -16,7 +16,7 @@ const { pipeline } = require('node:stream/promises');
 /**
  * transfer action
  */
-module.exports = async (fiber) => {
+module.exports = exports = async (fiber) => {
   logger.info("=== transfer");
   let retCode = 0;